Add files via upload
Added a timeout to the findSolution function so that the function will not run for more than 10 minutes.
Also, made sure that the queue would only add puzzles with a cost less than or equal to 31. This will make sure that the function will report failure if there is no solution.